fxing bugs and writing tests
