+++ collecte/data/orbis_d4g_sample.csv
@@ -1,0 +1,87 @@
+name,TIN,country_iso_code,name_normalized
+AGL ENERGY LIMITED,115061375,AU,agl energy limited
+ALLIANZ SE,FEI1007380,DE,allianz se
+BERLINER WASSERBETRIEBE,2010208815,DE,berliner wasserbetriebe
+BHP GROUP LIMITED,004028077,AU,bhp group limited
+BPOST SA,0214596464,BE,bpost sa
+BRENNTAG SE,5250167797,DE,brenntag se
+CSL LIMITED,051588348,AU,csl limited
+FRESENIUS MANAGEMENT SE,6290263710,DE,fresenius management se
+ORICA LIMITED,004145868,AU,orica limited
+SILTRONIC AG,8170662128,DE,siltronic ag
+SMA SOLAR TECHNOLOGY AG,6150026135,DE,sma solar technology ag
+SOUTH32 LIMITED,093732597,AU,south32 limited
+UNIPER SE,8350117034,DE,uniper se
+WESFARMERS LIMITED,008984049,AU,wesfarmers limited
+ADECCO GROUP AG,CHE107031232,CH,adecco group ag
+BANCO VOTORANTIM SA,59588111000103,BR,banco votorantim sa
+CARDNO LIMITED,108112303,AU,cardno limited
+IBERDROLA SA,A48010615,ES,iberdrola sa
+MAPFRE SA,A08055741,ES,mapfre sa
+ORSTED A/S,36213728,DK,orsted a/s
+REPSOL S.A,A78374725,ES,repsol s.a
+TECK RESOURCES LIMITED,20121NC,CA,teck resources limited
+TELEFONICA SA,A28015865,ES,telefonica sa
+TENCENT HOLDINGS LIMITED,30932FH,KY,tencent holdings limited
+ACCIONA SA,A08001851,ES,acciona sa
+ACERINOX SA,A28250777,ES,acerinox sa
+"AMADEUS IT GROUP, S.A.",A84236934,ES,"amadeus it group, s.a."
+"APPLUS SERVICES, S.A.",A64622970,ES,"applus services, s.a."
+COLOPLAST A/S,69749917,DK,coloplast a/s
+FERROVIAL SA,A81939209,ES,ferrovial sa
+INDRA SISTEMAS S.A.,A28599033,ES,indra sistemas s.a.
+RED ELECTRICA CORPORACION SA,A78003662,ES,red electrica corporacion sa
+TRYGHEDSGRUPPEN SMBA,10430410,DK,tryghedsgruppen smba
+ENAGAS SA,A28294726,ES,enagas sa
+MAXAMCORP HOLDING SL,B84598754,ES,maxamcorp holding sl
+ROYAL UNIBREW A/S,41956712,DK,royal unibrew a/s
+ANGLO AMERICAN PLC,03564138,GB,anglo american plc
+BT GROUP PLC,04190816,GB,bt group plc
+DIAGEO PLC,00023307,GB,diageo plc
+EVRAZ PLC,07784342,GB,evraz plc
+FORTUM OYJ,14636114,FI,fortum oyj
+LEGAL & GENERAL GROUP PLC,01417162,GB,legal & general group plc
+MACSF SGAM,488324617,FR,macsf sgam
+NATIONAL GRID PLC,04031152,GB,national grid plc
+PEARSON PLC,00053723,GB,pearson plc
+PRUDENTIAL PLC,01397169,GB,prudential plc
+RIO TINTO PLC,00719885,GB,rio tinto plc
+SCHRODERS PLC,03909886,GB,schroders plc
+SHELL PLC,04366849,NL,shell plc
+STANDARD CHARTERED PLC,00966425,GB,standard chartered plc
+TOTALENERGIES SE,542051180,FR,totalenergies se
+VODAFONE GROUP PLC,01833679,GB,vodafone group plc
+AJINOMOTO COMPANY INC,8010001034740,JP,ajinomoto company inc
+ATLANTIA S.P.A.,03731380261,IT,atlantia s.p.a.
+ENEL SPA,00811720580,IT,enel spa
+LEONARDO S.P.A.,00401990585,IT,leonardo s.p.a.
+OMRON CORPORATION,1130001016824,JP,omron corporation
+PRYSMIAN S.P.A.,04866320965,IT,prysmian s.p.a.
+RAKUTEN GROUP INC.,9010701020592,JP,rakuten group inc.
+SAIPEM SPA,00825790157,IT,saipem spa
+SUMITOMO CHEMICAL COMPANY LIMITED,2010001071327,JP,sumitomo chemical company limited
+ASTM S.P.A.,00488270018,IT,astm s.p.a.
+FERALPI HOLDING S.P.A.,01898120173,IT,feralpi holding s.p.a.
+SNAM S.P.A.,13271390158,IT,snam s.p.a.
+TERNA S.P.A.,05779661007,IT,terna s.p.a.
+AEGON NV,27076669,NL,aegon nv
+EQUINOR ASA,923609016,NO,equinor asa
+KONINKLIJKE KPN NV,02045200,NL,koninklijke kpn nv
+NORSK HYDRO ASA,914778271,NO,norsk hydro asa
+RANDSTAD N.V.,33216172,NL,randstad n.v.
+TELENOR ASA,982463718,NO,telenor asa
+TKH GROUP NV,06045666,NL,tkh group nv
+UNILEVER GROUP,91211EN,NL,unilever group
+YARA INTERNATIONAL ASA,986228608,NO,yara international asa
+HESS CORPORATION,134921002,US,hess corporation
+SWISS REINSURANCE GROUP,FEI1017131,US,swiss reinsurance group
+BARLOWORLD LIMITED,191800009506,ZA,barloworld limited
+BONAVA AB,5569280380,SE,bonava ab
+CIPLA LIMITED,30327FI,IN,cipla limited
+ENAV S.P.A.,97016000586,IT,enav s.p.a.
+ESSITY AKTIEBOLAG (PUBL),5563255511,SE,essity aktiebolag (publ)
+"FILA HOLDINGS CO.,LTD.",1101110792667,KR,"fila holdings co.,ltd."
+ADYEN N.V,34259528,NL,adyen n.v
+CELLNEX TELECOM S.A.,A64907306,ES,cellnex telecom s.a.
+"TIANJIN CHASE SUN PHARMACEUTICAL CO.,LTD",32049PC,CN,"tianjin chase sun pharmaceutical co.,ltd"
+"NTT TC LEASING CO.,LTD.",3010401151289,JP,"ntt tc leasing co.,ltd."
